NeatQueue: move types to separate file
